backend/requirements.txt
# Django and DRF
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
djoser==2.2.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Scraping tools
requests==2.31.0
beautifulsoup4==4.12.2
tabula-py==2.7.0
lxml==4.9.3

# Messaging
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
pika==1.3.2

# Geospatial libraries
django-leaflet==0.29.0
geojson==3.0.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.1.0
black==23.10.1
isort==5.12.0
flake8==6.1.0

# Documentation
drf-yasg==1.21.7
